x86: move CAR teardown to separate ramstage module
CAR teardown destroys the cache backing the current stage's stack and
BSS — it cannot run in the same stage that set up CAR. Move it to
car_teardown.rs, a separate module called at the start of ramstage
after the stack has moved to DRAM.

car.rs (setup, bootblock):
  - Remove _car_teardown assembly and car_teardown() wrapper
  - Add comment pointing to car_teardown.rs

car_teardown.rs (teardown, ramstage):
  - 64-bit assembly (ramstage runs in long mode with DRAM stack)
  - WBINVD to flush dirty CAR lines before disabling
  - Auto-detects NEM via MSR 0x2E0 and clears it
  - Clears MTRR0 (the CAR region) so it doesn't shadow DRAM
  - Re-enables MTRRs and cache for normal DRAM operation
  - Normal call/ret convention (has a real stack in DRAM)
